Agregacion de Imagenes A Productos

+++ docs/scripts/02_carretillas.sql
@@ -22,15 +22,16 @@ CREATE TABLE
         CONSTRAINT `carretilla_user_key` FOREIGN KEY (`usercod`) REFERENCES `usuario` (`usercod`) ON DELETE NO ACTION ON UPDATE NO ACTION,
         CONSTRAINT `carretilla_prd_key` FOREIGN KEY (`productId`) REFERENCES `products` (`productId`) ON DELETE NO ACTION ON UPDATE NO ACTION
     );
-
-CREATE TABLE
-    `carretillaanon` (
-        `anoncod` varchar(128) NOT NULL,
-        `productId` bigint(18) NOT NULL,
-        `crrctd` int(5) NOT NULL,
-        `crrprc` decimal(12, 2) NOT NULL,
-        `crrfching` datetime NOT NULL,
-        PRIMARY KEY (`anoncod`, `productId`),
-        KEY `productId_idx` (`productId`),
-        CONSTRAINT `carretillaanon_prd_key` FOREIGN KEY (`productId`) REFERENCES `products` (`productId`) ON DELETE NO ACTION ON UPDATE NO ACTION
-    );
+CREATE TABLE `carretillaanon` (
+    `anoncod` varchar(128) NOT NULL,
+    `productId` int(11) NOT NULL,
+    `crrctd` int(5) NOT NULL,
+    `crrprc` decimal(12, 2) NOT NULL,
+    `crrfching` datetime NOT NULL,
+    PRIMARY KEY (`anoncod`, `productId`),
+    KEY `productId_idx` (`productId`),
+    CONSTRAINT `carretillaanon_prd_key` FOREIGN KEY (`productId`) 
+        REFERENCES `products` (`productId`) 
+        ON DELETE NO ACTION 
+        ON UPDATE NO ACTION
+);
